pyproject.toml
[project]
name = "boatmon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "aiodns>=3.2.0",
    "aiohttp>=3.10.11",
    "rpi-bme280>=0.2.4",
]
